requirements.txt
requests
python-dotenv
selenium